edited: University and Speciality entities

+++ backend/internal/entities/speciality.go
@@ -4,7 +4,7 @@ type Speciality struct {
 	Id             int
 	UniversityName string
 	Name           string
-	BudgetPoints   float64
-	ContractPoints float64
+	BudgetPoints   int
+	ContractPoints int
 	Cost           int
 }
